pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-buy123-vendor"
version = "0.1.0"
description = "MCP Server for Buy123 Vendor API — AI-callable tools via Model Context Protocol"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Asgard AI Platform", email = "dev@asgard.ai"},
]
dependencies = [
    "requests>=2.31.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
auth = [
    "playwright>=1.45.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]
[project.scripts]
mcp-buy123-vendor = "mcp_buy123_vendor.server:main"
mcp-buy123-vendor-login = "mcp_buy123_vendor.scripts.playwright_login:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

include = ["mcp_buy123_vendor*"]
[tool.pytest.ini_options]
testpaths = ["tests"]
